# local-shared-data/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/logs/upload": {
      "post": {
        "tags": [
          "logs"
        ],
        "summary": "Upload Log File",
        "operationId": "upload_log_file",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_logs-upload_log_file"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/logs/files": {
      "get": {
        "tags": [
          "logs"
        ],
        "summary": "List Log Files",
        "operationId": "list_log_files",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number"
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Page size",
              "default": 50,
              "title": "Size"
            },
            "description": "Page size"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_LogFileRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/logs/files/{file_id}/entries": {
      "get": {
        "tags": [
          "logs"
        ],
        "summary": "List Log Entries",
        "operationId": "list_log_entries",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "file_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "File Id"
            }
          },
          {
            "name": "only_anomalies",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Only Anomalies"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number"
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Page size",
              "default": 50,
              "title": "Size"
            },
            "description": "Page size"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_LogEntryRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/logs/files/{file_id}/summary": {
      "get": {
        "tags": [
          "logs"
        ],
        "summary": "Get Log File Summary",
        "operationId": "get_log_file_summary",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "file_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "File Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogFileSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/logs/files/{file_id}": {
      "delete": {
        "tags": [
          "logs"
        ],
        "summary": "Delete Log File",
        "operationId": "delete_log_file",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "file_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "File Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/logs/entries/{entry_id}/explain": {
      "post": {
        "tags": [
          "logs"
        ],
        "summary": "Explain Log Entry",
        "description": "LLM-backed per-entry verdict.\n\nReturns `{\"verdict\": \"anomaly\" | \"false_positive\" | \"normal\",\n          \"description\": \"...\",\n          \"source\": \"llm\" | \"fallback\"}`.\n\nAuthorizes via the parent log file's ownership.",
        "operationId": "explain_log_entry",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "entry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Entry Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_logs-upload_log_file": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          },
          "model": {
            "type": "string",
            "title": "Model",
            "default": "transformer"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_logs-upload_log_file"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LogEntryRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "log_file_id": {
            "type": "string",
            "format": "uuid",
            "title": "Log File Id"
          },
          "timestamp": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timestamp"
          },
          "source_ip": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Ip"
          },
          "user_agent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Agent"
          },
          "action": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Action"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "method": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Method"
          },
          "status_code": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status Code"
          },
          "bytes_sent": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bytes Sent"
          },
          "url_category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url Category"
          },
          "threat_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Threat Name"
          },
          "user_login": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Login"
          },
          "raw_line": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Line"
          },
          "is_anomaly": {
            "type": "boolean",
            "title": "Is Anomaly"
          },
          "anomaly_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Anomaly Score"
          },
          "anomaly_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Anomaly Reason"
          }
        },
        "type": "object",
        "required": [
          "id",
          "log_file_id",
          "is_anomaly"
        ],
        "title": "LogEntryRead"
      },
      "LogFileRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "uploaded_at": {
            "type": "string",
            "format": "date-time",
            "title": "Uploaded At"
          },
          "total_entries": {
            "type": "integer",
            "title": "Total Entries"
          },
          "anomaly_count": {
            "type": "integer",
            "title": "Anomaly Count"
          }
        },
        "type": "object",
        "required": [
          "id",
          "filename",
          "uploaded_at",
          "total_entries",
          "anomaly_count"
        ],
        "title": "LogFileRead"
      },
      "LogFileSummary": {
        "properties": {
          "file": {
            "$ref": "#/components/schemas/LogFileRead"
          },
          "timeline": {
            "items": {
              "$ref": "#/components/schemas/TimelineBucket"
            },
            "type": "array",
            "title": "Timeline"
          },
          "top_source_ips": {
            "items": {
              "$ref": "#/components/schemas/TopIP"
            },
            "type": "array",
            "title": "Top Source Ips"
          },
          "top_actions": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Top Actions"
          },
          "top_categories": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Top Categories"
          },
          "ai_explanation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Explanation"
          }
        },
        "type": "object",
        "required": [
          "file",
          "timeline",
          "top_source_ips",
          "top_actions",
          "top_categories"
        ],
        "title": "LogFileSummary"
      },
...
      "Page_LogEntryRead_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/LogEntryRead"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Total"
          },
          "page": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Page"
          },
          "size": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Size"
          },
          "pages": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Pages"
          }
        },
        "type": "object",
        "required": [
          "items",
          "page",
          "size"
        ],
        "title": "Page[LogEntryRead]"
      },
      "Page_LogFileRead_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/LogFileRead"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Total"
          },
          "page": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Page"
          },
          "size": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Size"
          },
          "pages": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Pages"
          }
        },
        "type": "object",
        "required": [
          "items",
          "page",
          "size"
        ],
        "title": "Page[LogFileRead]"
      },
      "TimelineBucket": {
        "properties": {
          "bucket": {
            "type": "string",
            "format": "date-time",
            "title": "Bucket"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "anomaly_count": {
            "type": "integer",
            "title": "Anomaly Count"
          }
        },
        "type": "object",
        "required": [
          "bucket",
          "count",
          "anomaly_count"
        ],
        "title": "TimelineBucket"
      },
      "TopIP": {
        "properties": {
          "source_ip": {
            "type": "string",
            "title": "Source Ip"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "anomaly_count": {
            "type": "integer",
            "title": "Anomaly Count"
          }
        },
        "type": "object",
        "required": [
          "source_ip",
          "count",
          "anomaly_count"
        ],
        "title": "TopIP"
      },
      "UploadResponse": {
        "properties": {
          "file": {
            "$ref": "#/components/schemas/LogFileRead"
          },
          "parsed": {
            "type": "integer",
            "title": "Parsed"
          },
          "skipped": {
            "type": "integer",
            "title": "Skipped"
          },
          "anomalies": {
            "type": "integer",
            "title": "Anomalies"
          }
        },
        "type": "object",
        "required": [
          "file",
          "parsed",
          "skipped",
          "anomalies"
        ],
        "title": "UploadResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "auth/jwt/login"
          }
        }
      }
    }
  }
}